Let release build get space_groups.txt from bundle
Previously the unit cell analysis was not reliably usable by the app
bundle, since it required a specific file inside a specific directory.
This file (space_groups.txt) is now placed inside the Resources folder
inside the app bundle. The changes introduced in this commit, make it
possible for the app to access this file.
